Overhauled the database and the search functions.
Changed the game selector and searchbar to to be in a form.
Now, the user can edit both fields and the system will wait until the search button is clicked.
In addition, the search function will split the text into several pieces and search each piece separately, allowing for more accurate searches.